src/test/kotlin/com/tobert/cube/controllers/DraftControllerTest.kt
package com.tobert.cube.controllers

import com.nhaarman.mockito_kotlin.verify
import com.tobert.cube.services.DraftService
import org.junit.Test
import org.junit.runner.RunWith
import org.springframework.beans.factory.annotation.Autowired
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
import org.springframework.boot.test.mock.mockito.MockBean
import org.springframework.http.MediaType
import org.springframework.test.context.junit4.SpringRunner
import org.springframework.test.web.servlet.MockMvc
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders
import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status

@RunWith(SpringRunner::class)
@WebMvcTest(controllers = [DraftController::class])
class DraftControllerTest {
    @Autowired
    lateinit var mvc: MockMvc

    @MockBean
    lateinit var mockDraftService: DraftService

    @Test
    fun `it can start a draft`() {
        mvc.perform(MockMvcRequestBuilders.post("/draft/start")
                .contentType(MediaType.APPLICATION_FORM_URLENCODED)
        ).andExpect(status().is3xxRedirection)

        verify(mockDraftService).startDraft(15)
    }
}